requirements_cu10.txt
Cython==0.29.23
easydict==1.9
h5py==2.10.0
ipdb==0.13.9
json5==0.9.6
matplotlib==3.3.4
numpy==1.19.5
open3d==0.12.0
pandas==1.1.5
Pillow==8.4.0
plyfile==0.7.4
protobuf==3.20.0
PyYAML==6.0
pyglet==1.5.17
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
tensorboard==1.14.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==1.9
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorpack==0.9.0
torch==1.7.0
torchfile==0.1.0
torchvision==0.8.1
tqdm==4.64.0
transforms3d==0.3.1
trimesh==3.9.19